requirements.txt
flask
requests
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
torch>=1.7.0
pandas
ipython
psutil
torchvision
PyYAML>=5.3.1
tqdm>=4.64.0
matplotlib>=3.2.2
seaborn>=0.11.0
scipy>=1.4.1
gitpython

